+++ src/CliValidate.h
@@ -5,9 +5,9 @@ using namespace std;
 
 class CliValidate {
   public:
-    static bool axisInterval(float min, float max);
+    static bool axisInterval(double min, double max);
     static bool displaySize(int size);
-    static bool epsilon(float epsilon);
+    static bool epsilon(double epsilon);
     static bool maxDegree(int max);
     static bool polyCount(int count);
 };
