+++ src/Compute.cpp
@@ -6,8 +6,8 @@
 #include "Settings.h"
 using namespace std;
 
-	float Compute::execute(vector<float> polynomial, float x) {
-		float output = 0.0;
+	double Compute::execute(vector<double> polynomial, double x) {
+		double output = 0.0;
 		for (int i = polynomial.size()-1; i > -1; --i)
 		{ 
 			output += polynomial.at(i) * pow(x, i);
@@ -15,10 +15,10 @@ using namespace std;
 		return output;
 	};
 
-	int Compute::minimumIndexWithinYPlusEpsilon(vector<float> imageOfX, float y, Settings settings) {
+	int Compute::minimumIndexWithinYPlusEpsilon(vector<double> imageOfX, double y, Settings settings) {
 		for (int i = 0; i < imageOfX.size(); i++)
 		{	
-			float pointOnGraph = imageOfX.at(i); 
+			double pointOnGraph = imageOfX.at(i);
 			if (abs(pointOnGraph - y) < settings.epsilon)
 			{
 				return i;
@@ -27,17 +27,17 @@ using namespace std;
 		return -1;
 	};
 
-	vector<vector<float> > Compute::computeImageArray(vector<vector<float> > polynomialArray,
-                                                    vector<vector<float> > imageArray,
+	vector<vector<double> > Compute::computeImageArray(vector<vector<double> > polynomialArray,
+                                                    vector<vector<double> > imageArray,
                                                     Settings settings) {
-    float stepWidth = Settings::stepWidth(settings);
+    double stepWidth = Settings::stepWidth(settings);
 		for (int i = 0; i < polynomialArray.size(); i++)
 		{
-			vector<float> thisPolynomial = polynomialArray.at(i);
+			vector<double> thisPolynomial = polynomialArray.at(i);
       int xStepCount = Settings::xStepCount(settings);
 			for (int j = 0; j <= xStepCount; j++)
 			{
-				float x = settings.xMin + j * stepWidth; // j = (x - xMin)/stepWidth
+				double x = settings.xMin + j * stepWidth; // j = (x - xMin)/stepWidth
 				imageArray.at(i).at(j) = execute(thisPolynomial, x);		
 			}
 		}
