+++ src/Compute.h
@@ -7,9 +7,9 @@ using namespace std;
 class Compute {
 
 	public:
-		static float execute(vector<float> polynomial, float x);
-		static vector<vector<float> > computeImageArray(vector<vector<float> > polynomialArray, vector<vector<float> > imageArray, Settings settings);
-		static int minimumIndexWithinYPlusEpsilon(vector<float> imageOfX, float y, Settings settings);
+		static double execute(vector<double> polynomial, double x);
+		static vector<vector<double> > computeImageArray(vector<vector<double> > polynomialArray, vector<vector<double> > imageArray, Settings settings);
+		static int minimumIndexWithinYPlusEpsilon(vector<double> imageOfX, double y, Settings settings);
 };
 
 #endif
