+++ src/Draw.h
@@ -6,9 +6,9 @@ using namespace std;
 
 class Draw {
 	public:
-		static std::string createRow(vector<vector<float> > imageArray, float y, Settings settings);
-		static void render(vector<vector<float> > polynomialArray, Settings settings);
-		static char determineCharacterToRender(vector<float> imageOfX, float x, float y, Settings settings);
+		static std::string createRow(vector<vector<double> > imageArray, double y, Settings settings);
+		static void render(vector<vector<double> > polynomialArray, Settings settings);
+		static char determineCharacterToRender(vector<double> imageOfX, double x, double y, Settings settings);
 };
 
 #endif
